Split up query and registration tests
